Extract the duplicated logic to the external layer
